Fundamentals of Python Programming
This notebook focuses on the fundamentals of Python programming. It covers various topics such as syntax, functions, modules, complex data types, exceptions, program control flow, and more. In addition, it explores the use of Booleans, sorting elements of a list or dictionary, comprehension of lists, annotations, and assert statements. Whether you're a beginner or an experienced programmer, this notebook provides a comprehensive overview of Python programming, including practical examples and explanations of each topic.