Add files via upload
This Jupyter notebook provides a comprehensive introduction to algorithmic complexity, complexity theory, and various algorithms implemented over Python lists and numpy arrays. It covers the basics of searching algorithms and their complexity, along with practical implementations of different search techniques like naive search, search with the enumerate function, and search with the index method of the list.

Moreover, the notebook demonstrates the use of recursion algorithms, data structures, and various functions from numpy, such as generating arrays, finding elements with the where method, and using iterators. The dichotomous search algorithm is also implemented and measured to demonstrate the representation and implementation of algorithmic complexity.

Furthermore, the notebook provides an in-depth analysis of sorting algorithms like bubble sort and insertion sort and their complexities. It also covers some tricks that can be used to optimize algorithms and enhance their performance.

Finally, the notebook concludes with a comparative study of complexity in time to analyze the efficiency of different algorithms. The notebook is ideal for anyone who wants to learn about algorithmic complexity and its applications in real-world scenarios.